Fix over-active filters and over-aggressive javascript insertion
Many actions and scripts are being applied to all post types, with
negative side effects.

The main fixes are:
 - error checking code to avoid errors on multisite where a badge post
is not found on the current site
 - avoid applying filters where not needed
 - avoid loading scripts where not necessary

Some of this fixes should be considered hacks since some of them
completely disable functionality which is harmful / not correctly
implemented  and not strictly necessary to the current use case.

This checkin includes coverage of:
https://wordpress.org/support/topic/fix-content-filtering-is-over-aggres
sive/
https://wordpress.org/support/topic/code-incorrectly-runs-for-every-post
-type-with-errors-fix/